keyframe animations
* added keyframe-based animations
* applied directly to the model renderer
* moved material data structure from outside the model since materials are independent of models

+++ src/main/java/cinnamon/utils/Timer.java
@@ -14,7 +14,7 @@ public class Timer {
 
     public int update() {
         long CurrMs = getTime();
-        float diffMs = (float) (CurrMs - lastMs);
+        long diffMs = CurrMs - lastMs;
 
         tickDelta = diffMs / tickMs;
         lastMs = CurrMs;
